improve code coverage

+++ mlprodict/onnxrt/ops_cpu/op_constant.py
@@ -69,13 +69,13 @@ class Constant_12(OpRun):
         if hasattr(self, 'sparse_value') and self.sparse_value is not None:
             self.cst = self.sparse_value
         elif hasattr(self, 'value_float') and self.value_float is not None:
-            self.cst = self.value_float
+            self.cst = self.value_float.astype(numpy.float32)
         elif hasattr(self, 'value_floats') and self.value_floats is not None:
-            self.cst = self.value_floats
+            self.cst = self.value_floats.astype(numpy.float32)
         elif hasattr(self, 'value_int') and self.value_int is not None:
-            self.cst = self.value_int
+            self.cst = self.value_int.astype(numpy.int64)
         elif hasattr(self, 'value_ints') and self.value_ints is not None:
-            self.cst = self.value_ints
+            self.cst = self.value_ints.astype(numpy.int64)
         elif hasattr(self, 'value_string') and self.value_string is not None:
             self.cst = self.value_string
         elif hasattr(self, 'value_strings') and self.value_strings is not None:
